Step 2: Add a plugin
Each extension can contain one or more plugins, which are the ones that
add features to CKAN. Plugins are classes that always inherit from the
`SingletonPlugin` class.

Plugins are registered on the `setup.py` file, on the `[ckan.plugins]`
section:

    intro_plugin=ckanext.intro.plugin:ExampleIntroPlugin

The first part before the equal sign is the name of the plugin, which
you'll need to add to the `ckan.plugins` section of your CKAN
configuration file:

    ckan.plugins = intro_plugin

*Important!* Every time that you add a plugin to the `setup.py` file you
need to reinstall your extension. To do so, run the following from your
extension directory (remember to have your virtualenv activated!):

    python setup.py develop

If you don't do this you will get a `PluginNotFoundException` when
starting CKAN.