Step 3: Implement an interface
Plugins can interact with CKAN by implementing one of its *interfaces*.
These provide methods which are hooks into the internal CKAN logic,
allowing plugins to customize the default behaviour. For a full list of
the available interfaces, check

http://docs.ckan.org/en/latest/extensions/plugin-interfaces.html

or the source of the `ckan/plugins/interfaces.py` file.

To implement an Interface, plugins need to call the `implements` method
with the interface class as a parameter.

In our particular case, we are implementing the `IConfigurer` interface,
which allows to access and modify the CKAN config object. We are not
really doing yet, other than adding a log message with the currently
used plugins.

The `IConfigurer` interface is really useful because it is called right
at the start of the CKAN initialization, so you if your extension needs
to perform some action when CKAN is started, this is the method to use.