Step 4: The plugins toolkit and checking the CKAN version
Whenever possible, when calling any CKAN core functionality, this should
be done via the plugins toolkit. This is a set of classes and functions that
CKAN offers to plugins and that are garanteed to be stable between CKAN
releases. Using them will make easier that your extension works between
different CKAN versions.

You can see all the available classes and functions from the toolkit
here:

http://docs.ckan.org/en/latest/toolkit.html

On this particular step we are calling one of the functions from the
toolkit that allows to check the CKAN core version you are running on:
`check_ckan_version`. This allows to set things up differently depending on
the CKAN version, or to stop execution altogether if your extension is
not compatible with the CKAN version.

To learn more about CKAN versions and the release process, visit:

http://docs.ckan.org/en/latest/upgrading.html#ckan-releases