Step 5: Templates and snippets 1
All CKAN pages are rendered from *templates* (CKAN uses Jinja2 as the template
engine) Each page that you see in the CKAN frontend has template defined on
the `ckan/templates` folder (eg the home page is rendered from
`ckan/templates/home/index.html`). Not all the code used to render a particular
page is located in the same template file though, as templates *extend* parent
files which contain common elements and can call other files which are called
*snippets*. Snippets are useful to structure different parts of the page and
hold markup that is used multiple times.

For example the template for the search page
(`ckan/templates/package/search.html`) extends the template `page.html`, which
defines the general layout of the site (header, footer, etc) and calls snippets
for the search form (`snippets/search_form.html`) and the list of results
(`snippets/package_list.html`).

Each template is divided on a series of blocks:

    {% block block_name %}

        # content

    {% endblock %}

These blocks are the ones that extensions can use to extend or modify the core
CKAN templates.

On this step we add some small customizations to the footer of our site.

First of all notice how on `plugin.py` we use the plugins toolkit to call a
function that allows us to register our own template folder on our extension.
Any templates that are present in this folder will now override the ones on
CKAN core.

We've created our templates dir in `ckanext/intro/theme/templates`.

Have a look first of all at the footer template
`ckanext/intro/theme/templates/footer.html`. First thing we added was a
`{% ckan_extends %}` tag on top. This means that we want to extend the existing
core footer template. If we wanted to *completely replace* the core one with
our own, we wouldn't include it, but you generally will want to extend them, as
it will save you a lot of code.

After these there are two blocks, which match existing blocks on the core
footer template.

We modified the left column of the footer, adding an image to it and the
central column, adding some links. There is a crucial difference between the
two though, on the left column we completely *replaced* all the core content with
our image, while on the center column we *appended* content to the existing
one.

This is because on the second block we used the `{{ super() }}` tag. This
basically gets replaced with the contents of the core parent block.

On the next step we will see a bit more about templates and snippets.

*Note*: for a proper introduction to theming and templates have a look at the
theming tutorial:

http://docs.ckan.org/en/latest/theming.html

(At the time of writing this is still a work in progress, you can have a look
 at the draft docs here:

 http://docs.ckan.org/en/847-new-theming-docs/theming.html
 )