Step 6: Templates and snippets 2, variables and helpers
This steps adds two small customizations, one to the left sidebar of the
dataset page and one to the items of the dataset listing in the search
page.

As we registered our template path on the last step, we only need to
drop the relevant template files on the same path as the ones on CKAN
core and the templates will get used.

On the dataset page template
(`ckanext/intro/theme/templates/package/read_base.html`) we extend the
`secondary_content` block adding an extra section. Again, notice the
`{{ super() }}` block that we use. If you remove it, our custom section
will be the only one displayed.

Templates have variables available that allow them render data or call
helper functions. Some of them are available to all templates. Here you
can find a list of them and what they are used for:

http://docs.ckan.org/en/latest/theming/variables-and-functions.html

(Use this if the above link is not yet available:
 http://docs.ckan.org/en/847-new-theming-docs/theming/variables-and-functions.html
)

Other variables are passed from the controllers to specific template,
like for instance the dataset being rendered on the dataset page (called
`package_dict` or `pkg`).

On our dataset page template we are calling `h`, which contains helper
functions for templates, and also using `pkg`, the variable that
contains all fields of the dataset being rendered.

We've also modified the search listing to display some extra fields from
the dataset. To do so we have modified the snippet that renders each
dataset item
(`ckanext/intro/theme/templates/snippets/package_item.html`).

Sadly snippets can not be extended at the moment, so if you want to
modify an existing core snippet you have to copy all its contents and
modify them.

In our case, we added some extra fields using the dataset variable (in
this case called `package`) and a helper function from `h`.

You can of course write your own snippets and call them from your
extension templates. Just be careful to not name them in the same way of
one of the existing core snippets.