Step 7: The public directory and custom routes
On this step we created a completely custom page (ie we didn't extend
one of the exsting ones in CKAN) which displays an image packaged with
our extension.

To be able to serve resources like images, Javascript or CSS files from
your own extension you need to register a directory as a public folder.
This is done with the `toolkit.add_public_directory` function, similarly
to how we did it for the extensions.

Once registered, files on the directory will be served from the CKAN
root level, for example:

http://yourckaninstance/images/sloth.jpg

We also mapped a custom URL (`/custom`) to a specific controllers, which
renders our custom template. This is done using the `IRoutes` interface:

http://docs.ckan.org/en/latest/extensions/plugin-interfaces.html#ckan.plugins.interfaces.IRoutes

(Note how we passed a keyword `inherit=True` when calling the
`implements` method. This is because this interface has other methods
that we are not going to use, and we are happy to use the default
version of them. If we don't use the `inherit` we will need to add all
methods to our plugin to avoid an exception)

Routes map specific URLs (like `/dataset` or `/group/<id>`) to
controllers and actions that will perform some tasks and render the
relevant template to diplay on the frontend.

We added a small controller on our plugin file which just renders our
custom template. Note how all the functions and classes used from CKAN
core are imported via the plugins toolkit.

To make easier to access our page, we added a link to our custom footer
template.