Step 8: Action and Authorization functions
A quick introduction to Action and Authorization functions. Every action that
can be performed on the web frontend or the API is implemented by an
*action function*. These are located in the `create.py`, `update.py`,
`delete.py` and `get.py` files in the `ckan/logic/action` folder. Regardless of
the interface calling the action (web, API, command line) the same function is
called. To see a list of all available actions visit:

http://docs.ckan.org/en/latest/api.html#api-reference

For each action function, there is an *authorization function*, which is called
on each action to see whether the user is authorized to perform it. The default
authorization functions are located in the `ckan/logic/auth` folder, with the
same file and function names as their actions.

Extensions can modify this default authorization function or add new ones for
their own actions implementing the `IAuthFunctions` interface.

To learn more about action and authorization functions, it is very recommended
that you read the following section of the tutorial for writing CKAN
extensions:

http://docs.ckan.org/en/latest/extensions/tutorial.html#implementing-the-iauthfunctions-plugin-interface

In our case, we have modified the default authorization to prevent users from
creating groups. Note how if you are logged in as a normal user (not a
sysadmin) you don't see the "Add Group" button on the groups page.

To do so, our plugin implements `IAuthFunctions` and adds a
`get_auth_function` methods that returns a mapping of our own auth functions.

The auth function that we need to modify is `group_create`. Auth functions need
to return a dict with a `success` key set to True if the user is authorized to
perform the action, or `False` otherwise. In our case then, we just return

   {'success': False}

You could of course, perform much more complex checks. Auth functions always
get two dicts when called: `context` and `data_dict`. `context` has
information about the request, like the user name, and gives access to the CKAN
model, which allows to query the database. `data_dict` is the data posted by
the user on the request. In our case they are the details of the group being
created.

Note: Authorization does not affect sysadmins. Sysadmins can perform any action
so there is no need to consider them on auth functions.

The tutorial for writing extensions contains a more complex example of a custom
authoriation function:

http://docs.ckan.org/en/latest/extensions/tutorial.html