Step 9: Action and Auhtorization functions 2: custom actions
On the last step we saw what action and authorization functions are and how you
register them from your plugin.

On this step, we added a new route (see Step 7), `/csv`, that when visited
outputs a CSV file with a report of all the datasets present on your CKAN
instance. The route points to a new method on our small controller, which in
turn calls an action function called `datasets_report_csv`.

This action function is defined on our `plugin.py` file, and we have registered
it using the `get_actions` method of the `IActions` interface. As with all
action and auth functions, receives a context and data_dict objects, although
in our case we won't use them.

The first thing that the action function does is check if the user is
authorized, this is done calling `toolkit.check_access` with the name of the
registered auth function. If one of your action functions needs the same
authorization as an existing action function you can of course call it as well.
For instance if you want to allow users to do some action on a dataset only if
they can edit it, you would call:

    p.toolkit.check_access('package_update', context, data_dict)

Note how our controller captures the `toolkit.NotAuhtorized` exceptions raised
by the authorization functions and aborts the request with a 401 status.

If the user is authorized (ie on our case is a sysadmin) we then call the
`package_search` action to get details of all datasets, and use the python
`csv` module to create a CSV file. The path of this file is passed to the
controller, which will read it and output it.

Note how we modify the headers of the response to reflect that we are outputing
a CSV file.

To request the dataset report, we added a link to the footer as well.